requirements.txt
aiogram~=3.21.0
pydantic~=2.11.7
requests~=2.32.4
mysql-connector-python
psutil